requirements.txt
Django==1.9
django-cors-headers==1.1.0
django-dotenv==1.3.0
djangorestframework==3.3.1
factory-boy==2.6.0
fake-factory==0.5.3
wheel==0.24.0
